refactor: Move utils.columns into utils.common.columns
