+++ utils/common/README.md
@@ -1,0 +1,6 @@
+# `utils.common`
+
+This directory contains common utility code, mainly little single-purpose helper functions.
+
+You most likely won't need to look at the code in this directory,
+except when you want to add more utility functions.
