+++ utils/common/columns.py
@@ -1,4 +1,6 @@
+"""Utility functions for reordering columns in polars dataframes."""
 import polars as pl
+
 
 def pull_columns_to_front(*columns: str) -> list[pl.Expr]:
     return [
